Changed controllers to use Response objects, thyme
Changed controllers to use Reposnse objects, which also required to
change tests a little bit.

Made some visual changes to thymeleaf templates, now its possible to see
all the flashcards from the set, but it still isnt working correctly,
this is work in progress.

Made change to Security Config to make it impossible to create a set
without being authenticated, cause it was too confussing.