Flashcards works, added active and recent sets
Its now possible to actually play with flashcards. If a user begins to
learn from a set, it will remember which cards he swiped as remembered,
so even if they close a set and reopen it, only unremembered cards will
show. The set will reload, which means that all cards from the set will
be shown, after some predefined time, which increaces as the user
completes the same set again.

When the user completes a set, it will be reloaded automatically, or
they can do it on their own by clicking the reload button, so they can
run a set again if they want to.

There is also functionality of showing to user their 5 most recent sets
that they have seen, so its easier for them to learn from those sets
again.

Also created a Bootstrap.java class which will populate database with
some test data, so its easier to test the application in browser.