Changed structure of handling requests to sets
Right now those requests to the flashcards sets which are connected with
UsersActiveSet, so those related to learning a set, are being handled by
UsersActiveSet -Api and -Service so it makes more sense.

Also right now its possible to just have a look at a set, without
learning from it.